// crm-frontend/lib/types/dashboard.ts
export interface ProvisionTransaction {
  date: string;
  type: string;
  ref: string;
  closeLeadId: string;
  amount: number;
  oppValue?: number;
  setupFee?: number;
  monthlyValue?: number;
  contractDuration?: number;
  source?: string;
  contractUrl?: string;
  empfehlungBonus?: number;
}